Add unit tests for stripped, not available and available PDBs
