+++ src/Shared/SharedAssemblyInfo.cs
@@ -22,6 +22,8 @@ using System.Runtime.InteropServices;
 
 [assembly: InternalsVisibleTo("BinSkim")]
 [assembly: InternalsVisibleTo("UnitTests.BinSkim.Rules")]
+[assembly: InternalsVisibleTo("Test.UnitTests.BinaryParsers")]
+[assembly: InternalsVisibleTo("Test.FunctionalTests.BinSkim.Driver")]
 [assembly: InternalsVisibleTo("BinSkim.Driver.FunctionalTests")]
 [assembly: InternalsVisibleTo("BinSkim.Rules.FunctionalTests")]
